backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-countries
Faker
tqdm
